+++ CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed cross-thread stack overflow false positives in parallel mode by updating stack baseline before QuickJS C entry points
 - Fixed iterators to use correct IteratorPrototype chain
 
+### Added
+
+- Add inheritance support for Rust classes #[#587](https://github.com/DelSkayn/rquickjs/pull/587)
+
 ## [0.11.0] - 2025-12-16
 
 ### Added